pyproject.toml
[project]
name = "dario-orch"
version = "1.0.0"
description = "DARIO Orchestrator Runtime — Self-Evolving AI Agent OS"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "psycopg[binary,pool]>=3.2.0",
    "pydantic-settings>=2.5.0",
    "httpx>=0.27.0",
    "ruamel.yaml>=0.18.0",
    "apscheduler>=3.10.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "httpx"]
